backend/requirements.txt
fastapi>=0.100.0,<0.111.0
uvicorn[standard]>=0.22.0,<0.29.0
sqlalchemy>=2.0.0,<2.1.0
pydantic[email]>=2.0.0,<2.7.0
pydantic-settings>=2.0.0,<2.2.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.7,<0.0.9
httpx>=0.24.0,<0.28.0
pyyaml>=6.0,<6.1
loguru>=0.7.0,<0.8.0
alembic>=1.10.0,<1.14.0
psycopg2-binary>=2.9.0,<2.10.0 # For PostgreSQL support, optional if only using SQLite but good for Alembic
gitpython>=3.1.0,<3.2.0
pytest>=7.0.0,<8.2.0
black>=23.0.0,<24.4.0
python-gedcom>=3.0.0,<3.1.0 # For parsing GEDCOM files
beautifulsoup4>=4.12.0,<4.13.0 # For web scraping by tools
ratelimit>=2.2.0,<2.3.0 # For rate limiting external API calls